- Fixes MLResultData.kt and adds MLListData.kt to resolve get issues. - Recycler view ready. - Adds some useful utils

+++ app/src/main/java/cl/maleb/mercadolibre/challenge/api/list/Cancellations.kt
@@ -2,6 +2,6 @@ package cl.maleb.mercadolibre.challenge.api.list
 
 data class Cancellations(
     val period: String? = null,
-    val rate: Int? = null,
+    val rate: Double? = null,
     val value: Int? = null
 )
