+++ app/src/main/java/cl/maleb/mercadolibre/challenge/api/list/Installments.kt
@@ -4,5 +4,5 @@ data class Installments(
     val amount: Double? = null,
     val currency_id: String? = null,
     val quantity: Int? = null,
-    val rate: Int? = null
+    val rate: Double? = null
 )
